[#146]test: 마이페이지 액션 컴포넌트 테스트 추가 및 renderWithClient 유틸 분리
